refactor(rpc): move directory
