+++ pkg/rpc/client/library/client_test.go
