+++ pkg/rpc/client/ospkg/client_test.go
