+++ pkg/rpc/server/library/server_test.go
