+++ pkg/rpc/server/ospkg/server.go
@@ -1,4 +1,4 @@
-package library
+package ospkg
 
 import (
 	"context"
@@ -6,9 +6,9 @@ import (
 	"github.com/google/wire"
 	"golang.org/x/xerrors"
 
-	"github.com/aquasecurity/trivy/internal/rpc"
-	detector "github.com/aquasecurity/trivy/pkg/detector/library"
+	detector "github.com/aquasecurity/trivy/pkg/detector/ospkg"
 	"github.com/aquasecurity/trivy/pkg/log"
+	"github.com/aquasecurity/trivy/pkg/rpc"
 	"github.com/aquasecurity/trivy/pkg/vulnerability"
 	proto "github.com/aquasecurity/trivy/rpc/detector"
 )
@@ -28,14 +28,14 @@ func NewServer(detector detector.Operation, vulnClient vulnerability.Operation) 
 	return &Server{detector: detector, vulnClient: vulnClient}
 }
 
-func (s *Server) Detect(ctx context.Context, req *proto.LibDetectRequest) (res *proto.DetectResponse, err error) {
-	vulns, err := s.detector.Detect(req.FilePath, rpc.ConvertFromRpcLibraries(req.Libraries))
+func (s *Server) Detect(ctx context.Context, req *proto.OSDetectRequest) (res *proto.DetectResponse, err error) {
+	vulns, eosl, err := s.detector.Detect(req.OsFamily, req.OsName, rpc.ConvertFromRpcPkgs(req.Packages))
 	if err != nil {
 		log.Logger.Warn(err)
-		return nil, xerrors.Errorf("failed to detect library vulnerabilities: %w", err)
+		return nil, xerrors.Errorf("failed to detect OS package vulnerabilities: %w", err)
 	}
 
 	s.vulnClient.FillInfo(vulns, false)
 
-	return &proto.DetectResponse{Vulnerabilities: rpc.ConvertToRpcVulns(vulns)}, nil
+	return &proto.DetectResponse{Vulnerabilities: rpc.ConvertToRpcVulns(vulns), Eosl: eosl}, nil
 }
