+++ pkg/rpc/server/ospkg/server_test.go
